refactor(FIR-35470): remove redundant account id resolution
