+++ src/connection/connection_v2.ts
@@ -5,7 +5,7 @@ import {
   QUERY_URL
 } from "../common/api";
 
-import { Connection as BaseConnection, AccountInfo } from "./base";
+import { Connection as BaseConnection } from "./base";
 import { Cache, inMemoryCache, noneCache } from "../common/tokenCache";
 
 export class ConnectionV2 extends BaseConnection {
@@ -56,36 +56,6 @@ export class ConnectionV2 extends BaseConnection {
     }
   }
 
-  async resolveAccountInfo(): Promise<AccountInfo> {
-    if (this.accountInfo === undefined) {
-      const cachedValue = this.cache.accountInfoStorage.get({
-        account: this.account,
-        apiEndpoint: this.context.apiEndpoint
-      });
-      if (cachedValue) {
-        this.accountInfo = cachedValue;
-      } else {
-        const { httpClient, apiEndpoint } = this.context;
-        const url = `${apiEndpoint}/${ACCOUNT_ID_BY_NAME(this.account)}`;
-        const { id, infraVersion } = await httpClient
-          .request<{ id: string; region: string; infraVersion: string }>(
-            "GET",
-            url
-          )
-          .ready();
-        this.accountInfo = { id, infraVersion: parseInt(infraVersion ?? "2") };
-        this.cache.accountInfoStorage.set(
-          {
-            account: this.account,
-            apiEndpoint: this.context.apiEndpoint
-          },
-          this.accountInfo
-        );
-      }
-    }
-    return this.accountInfo;
-  }
-
   async resolveEngineEndpoint() {
     const { engineName, database } = this.options;
     // Connect to system engine first
@@ -93,7 +63,6 @@ export class ConnectionV2 extends BaseConnection {
       await this.getSystemEngineEndpointAndParameters();
     this.engineEndpoint = new URL(QUERY_URL, systemUrl).href;
     this.parameters = { ...this.parameters, ...systemParameters };
-    this.accountInfo = await this.resolveAccountInfo();
 
     if (database) {
       await this.execute(`USE DATABASE "${database}"`);
