+++ src/service/index.ts
@@ -4,7 +4,7 @@ import { DatabaseService as DatabaseServiceV1 } from "./database/v1";
 import { EngineServiceInterface } from "./engine/types";
 import { EngineService as EngineServiceV2 } from "./engine";
 import { EngineService as EngineServiceV1 } from "./engine/v1";
-import { ResourceManagerContext } from "../types";
+import { ResourceManagerContext, ResourceManagerContextV1 } from "../types";
 
 export class ResourceManager {
   private context: ResourceManagerContext;
@@ -18,8 +18,12 @@ export class ResourceManager {
       this.engine = new EngineServiceV2(this.context);
       this.database = new DatabaseServiceV2(this.context);
     } else if (httpClient.authenticator.isUsernamePassword()) {
-      this.engine = new EngineServiceV1(this.context);
-      this.database = new DatabaseServiceV1(this.context);
+      this.engine = new EngineServiceV1(
+        this.context as ResourceManagerContextV1
+      );
+      this.database = new DatabaseServiceV1(
+        this.context as ResourceManagerContextV1
+      );
     } else {
       throw new Error(
         "Invalid auth credentials provided. Please check your credentials and try again."
