+++ test/unit/v2/connection.test.ts
@@ -2,8 +2,6 @@ import { setupServer } from "msw/node";
 import { rest } from "msw";
 import { Firebolt } from "../../../src";
 import { ConnectionOptions } from "../../../src/types";
-import { AccountNotFoundError } from "../../../src/common/errors";
-import { ConnectionV2 } from "../../../src/connection/connection_v2";
 import { QUERY_URL } from "../../../src/common/api";
 import { inMemoryCache } from "../../../src/common/tokenCache";
 
@@ -25,32 +23,6 @@ const engineUrlResponse = {
     }
   ],
   data: [["https://some_engine.com", "dummy", "Running"]],
-  rows: 1
-};
-
-const selectOneResponse = {
-  meta: [
-    {
-      name: "one",
-      type: "Int32"
-    }
-  ],
-  data: [
-    {
-      one: 1
-    }
-  ],
-  rows: 1
-};
-
-const selectAttachedToResponse = {
-  meta: [
-    {
-      name: "attached_to",
-      type: "Text"
-    }
-  ],
-  data: [["dummy2"]],
   rows: 1
 };
 
@@ -84,149 +56,6 @@ describe("Connection V2", () => {
     });
   });
 
-  it("shows a helpful error message on account not found", async () => {
-    const firebolt = Firebolt({
-      apiEndpoint
-    });
-
-    server.use(
-      rest.get(
-        `https://api.fake.firebolt.io/web/v3/account/invalid_account/resolve`,
-        (req, res, ctx) => {
-          return res(ctx.status(404));
-        }
-      ),
-      rest.get(
-        `https://api.fake.firebolt.io/web/v3/account/invalid_account/engineUrl`,
-        (req, res, ctx) => {
-          return res(ctx.status(404));
-        }
-      )
-    );
-
-    const connectionParams: ConnectionOptions = {
-      auth: {
-        client_id: "dummy",
-        client_secret: "dummy"
-      },
-      database: "dummy",
-      engineName: "dummy",
-      account: "invalid_account"
-    };
-
-    await expect(async () => {
-      await firebolt.testConnection(connectionParams);
-    }).rejects.toThrow(AccountNotFoundError);
-  });
-  it("retrieves account version", async () => {
-    const firebolt = Firebolt({
-      apiEndpoint
-    });
-
-    server.use(
-      rest.post(`https://id.fake.firebolt.io/oauth/token`, (req, res, ctx) => {
-        return res(
-          ctx.json({
-            access_token: "fake_access_token"
-          })
-        );
-      }),
-      rest.get(
-        `https://api.fake.firebolt.io/web/v3/account/my_account/resolve`,
-        (req, res, ctx) => {
-          return res(
-            ctx.json({
-              id: "123",
-              infraVersion: 2
-            })
-          );
-        }
-      ),
-      rest.get(
-        `https://api.fake.firebolt.io/web/v3/account/my_account/engineUrl`,
-        (req, res, ctx) => {
-          return res(
-            ctx.json({
-              engineUrl: "https://some_system_engine.com"
-            })
-          );
-        }
-      ),
-      rest.post(
-        `https://some_system_engine.com/${QUERY_URL}`,
-        (req, res, ctx) => {
-          return res(ctx.json(engineUrlResponse));
-        }
-      )
-    );
-
-    const connectionParams: ConnectionOptions = {
-      auth: {
-        client_id: "dummy",
-        client_secret: "dummy"
-      },
-      database: "dummy",
-      account: "my_account"
-    };
-
-    const connection = await firebolt.connect(connectionParams);
-    const accountInfo = await (connection as ConnectionV2).resolveAccountInfo();
-    expect(accountInfo.infraVersion).toBe(2);
-  });
-  it("defaults account version to 2 if not provided", async () => {
-    const firebolt = Firebolt({
-      apiEndpoint
-    });
-
-    server.use(
-      rest.post(`https://id.fake.firebolt.io/oauth/token`, (req, res, ctx) => {
-        return res(
-          ctx.json({
-            access_token: "fake_access_token"
-          })
-        );
-      }),
-      rest.get(
-        `https://api.fake.firebolt.io/web/v3/account/my_account/resolve`,
-        (req, res, ctx) => {
-          return res(
-            ctx.json({
-              id: "123"
-            })
-          );
-        }
-      ),
-      rest.get(
-        `https://api.fake.firebolt.io/web/v3/account/my_account/engineUrl`,
-        (req, res, ctx) => {
-          return res(
-            ctx.json({
-              engineUrl: "https://some_system_engine.com?param=value"
-            })
-          );
-        }
-      ),
-      rest.post(
-        `https://some_system_engine.com/${QUERY_URL}`,
-        (req, res, ctx) => {
-          return res(ctx.json(engineUrlResponse));
-        }
-      )
-    );
-
-    const connectionParams: ConnectionOptions = {
-      auth: {
-        client_id: "dummy",
-        client_secret: "dummy"
-      },
-      database: "dummy",
-      account: "my_account"
-    };
-
-    const connection = await firebolt.connect(connectionParams);
-    const accountInfo = await (connection as ConnectionV2).resolveAccountInfo();
-    expect(accountInfo.infraVersion).toBe(2);
-  });
   it("respects system engine query parameters for account version 2", async () => {
     let systemEngineParamsUsed = {};
     server.use(
@@ -237,17 +66,6 @@ describe("Connection V2", () => {
           })
         );
       }),
-      rest.get(
-        `https://api.fake.firebolt.io/web/v3/account/my_account/resolve`,
-        (req, res, ctx) => {
-          return res(
-            ctx.json({
-              id: "123",
-              infraVersion: 2
-            })
-          );
-        }
-      ),
       rest.get(
         `https://api.fake.firebolt.io/web/v3/account/my_account/engineUrl`,
         (req, res, ctx) => {
@@ -286,12 +104,11 @@ describe("Connection V2", () => {
     expect(systemEngineParamsUsed).toHaveProperty("param");
     expect(systemEngineParamsUsed).not.toHaveProperty("account_id");
   });
-  it("caches account info", async () => {
+  it("caches system engine info", async () => {
     const firebolt = Firebolt({
       apiEndpoint
     });
 
-    let resolveCalls = 0;
     let engineUrlCalls = 0;
 
     server.use(
@@ -302,18 +119,6 @@ describe("Connection V2", () => {
           })
         );
       }),
-      rest.get(
-        `https://api.fake.firebolt.io/web/v3/account/my_account/resolve`,
-        (req, res, ctx) => {
-          resolveCalls++;
-          return res(
-            ctx.json({
-              id: "123",
-              infraVersion: 2
-            })
-          );
-        }
-      ),
       rest.get(
         `https://api.fake.firebolt.io/web/v3/account/my_account/engineUrl`,
         (req, res, ctx) => {
@@ -351,17 +156,10 @@ describe("Connection V2", () => {
       account: "my_account"
     };
 
-    const connection = await firebolt.connect(connectionParams);
-    const accountInfo = await (connection as ConnectionV2).resolveAccountInfo();
+    await firebolt.connect(connectionParams);
 
-    const connection2 = await firebolt.connect(connectionParams);
-    const accountInfo2 = await (
-      connection2 as ConnectionV2
-    ).resolveAccountInfo();
+    await firebolt.connect(connectionParams);
 
-    expect(accountInfo).toStrictEqual(accountInfo2);
-
-    expect(resolveCalls).toBe(1);
     expect(engineUrlCalls).toBe(1);
   });
   it("respects useCache option", async () => {
@@ -369,7 +167,6 @@ describe("Connection V2", () => {
       apiEndpoint
     });
 
-    let resolveCalls = 0;
     let engineUrlCalls = 0;
 
     server.use(
@@ -380,18 +177,6 @@ describe("Connection V2", () => {
           })
         );
       }),
-      rest.get(
-        `https://api.fake.firebolt.io/web/v3/account/my_account/resolve`,
-        (req, res, ctx) => {
-          resolveCalls++;
-          return res(
-            ctx.json({
-              id: "123",
-              infraVersion: 2
-            })
-          );
-        }
-      ),
       rest.get(
         `https://api.fake.firebolt.io/web/v3/account/my_account/engineUrl`,
         (req, res, ctx) => {
@@ -421,13 +206,10 @@ describe("Connection V2", () => {
       useCache: false
     };
 
-    const connection = await firebolt.connect(connectionParams);
-    await (connection as ConnectionV2).resolveAccountInfo();
+    await firebolt.connect(connectionParams);
 
-    const connection2 = await firebolt.connect(connectionParams);
-    await (connection2 as ConnectionV2).resolveAccountInfo();
+    await firebolt.connect(connectionParams);
 
-    expect(resolveCalls).toBe(2);
     expect(engineUrlCalls).toBe(2);
   });
 });
